Address all compilation errors
