+++ crates/ibc-clients/ics07-tendermint/Cargo.toml
@@ -25,3 +25,16 @@ ibc-proto = { workspace = true, features = ["server"] }
 tendermint-light-client-verifier = { workspace = true, features = ["rust-crypto"] }
 
 prost = { workspace = true }
+
+[dev-dependencies]
+env_logger = { workspace = true }
+test-log = { workspace = true }
+tracing-subscriber = { workspace = true }
+
+[features]
+default = ["std"]
+
+std = [
+    "ibc-proto/std",
+    "prost/std",
+]
